requirements.txt
Django==3.1.3
Pillow==9.3.0
torch==1.12.1+cu116
torchvision==0.13.1+cu116
